Local instance - Makefile and test updated to work in splitted mode
