+++ local/run_local
@@ -1,2 +1,3 @@
-cd local
+make build
+cd src
 node start.js $1
