+++ local/tests/package.json
@@ -36,12 +36,9 @@
     "url": "http://www.gnu.org/licenses/agpl-3.0.html"
   },
   "scripts": {
-    "test": "npm run-script test_libs ; npm run-script test_local ; npm run-script test_global",
-    "test_libs": "./libs/unit/pre_tests.js ; ./node_modules/.bin/vows libs/unit/test_*.js -i --spec ; ./libs/unit/post_tests.js",
-    "test_local": "./node_modules/.bin/vows local/unit/test_*.js -i --spec ; ./node_modules/.bin/vows local/functional/test_*.js -i --spec",
-    "test_global": "./node_modules/.bin/vows global/unit/test_*.js -i --spec ; ./node_modules/.bin/vows global/functional/test_*.js -i --spec",
-    "start": "npm run-script start_local ; npm run-script start_global",
-    "start_local": "cd ../src/; ./run_local &",
-    "start_global": "cd ../src/; ./run_global &"
+    "test": "npm run-script test_unit ; npm run-script test_functional",
+    "test_unit": "./node_modules/.bin/vows unit/test_*.js -i --spec",
+    "test_functional": "./node_modules/.bin/vows functional/test_*.js -i --spec",
+    "start": "cd ../; ./run_local &"
   }
 }
