+++ local/tests/unit/test_sandmand_tcp.js
@@ -7,7 +7,7 @@
  */
 
 require('./configuration.js');
-var sandman_tcp = require('../../../src/local/modules/sandmans/tcp.js'),
+var sandman_tcp = require('../../src/modules/sandmans/tcp.js'),
     mock_tcpserver = require('../mocks/tcp_server_mock'),
     assert = require('assert'),
     vows = require('vows');
