+++ local/tests/unit/test_sandmand_udp.js
@@ -7,7 +7,7 @@
  */
 
 require('./configuration.js');
-var sandman_udp = require('../../../src/local/modules/sandmans/udp.js'),
+var sandman_udp = require('../../src/modules/sandmans/udp.js'),
     mock_udpserver = require('../mocks/udp_server_mock'),
     buffer = require('buffer'),
     assert = require('assert'),
